+++ src/BusinessLogic/ShippingMethod/Utility/ShipmentStatus.php
@@ -67,16 +67,4 @@ class ShipmentStatus
                 return self::STATUS_PENDING;
         }
     }
-
-    /**
-     * Returns whether shipment labels should be fetched.
-     *
-     * @param string $shipmentStatus Raw shipment status from Packlink.
-     *
-     * @return bool Returns TRUE if labels should be fetched; otherwise returns FALSE.
-     */
-    public static function shouldFetchLabels($shipmentStatus)
-    {
-        return in_array($shipmentStatus, array('READY_TO_PRINT', 'READY_FOR_COLLECTION', 'IN_TRANSIT', 'DELIVERED'));
-    }
 }