+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - `isRegistered` method in `RepositoryRegistry`.
+- `isShipmentLabelsAvailable` method in `OrderService`
 
 ### Removed